+++ app/inbound/save-invoice-lines.js
@@ -4,7 +4,11 @@ const saveInvoiceLines = async (invoiceLines, paymentRequestId, transaction) => 
   for (const invoiceLine of invoiceLines) {
     delete invoiceLine.invoiceLineId
     invoiceLine.fundingCode = invoiceLine.schemeCode
-    await db.invoiceLine.create({ ...invoiceLine, paymentRequestId }, { transaction })
+    try {
+      await db.invoiceLine.create({ ...invoiceLine, paymentRequestId }, { transaction })
+    } catch (err) {
+      // console.error(err)
+    }
   }
 }
